refactor: 뒤로가기 이벤트 리스너 커스텀 훅으로 분리
