+++ src/pages/interest/InterestPage.tsx
@@ -1,10 +1,9 @@
-import { graniteEvent } from "@apps-in-toss/web-framework";
 import { adaptive, colors } from "@toss/tds-colors";
 import { Checkbox, FixedBottomCTA, List, ListRow, Top } from "@toss/tds-mobile";
-import { useEffect } from "react";
 import { topics } from "../../constants/topics";
 import { useMultiStep } from "../../contexts/MultiStepContext";
 import { useSurvey } from "../../contexts/SurveyContext";
+import { useBackEventListener } from "../../hooks/useBackEventListener";
 import { createSurveyInterests } from "../../service/form";
 
 export const InterestPage = () => {
@@ -25,22 +24,13 @@ export const InterestPage = () => {
 		}
 	};
 
-	useEffect(() => {
-		const unsubscription = graniteEvent.addEventListener("backEvent", {
-			onEvent: () => {
-				if (state.screening.enabled) {
-					handleStepChange(2);
-				} else {
-					handleStepChange(1);
-				}
-			},
-			onError: (error) => {
-				alert(`에러가 발생했어요: ${error}`);
-			},
-		});
-
-		return unsubscription;
-	}, [handleStepChange, state.screening.enabled]);
+	useBackEventListener(() => {
+		if (state.screening.enabled) {
+			handleStepChange(2);
+		} else {
+			handleStepChange(1);
+		}
+	});
 
 	const handleNext = async () => {
 		const interests = selectedTopics.map((topic) => topic.value);
