+++ src/pages/screening/ScreeningOption.tsx
@@ -1,9 +1,8 @@
-import { graniteEvent } from "@apps-in-toss/web-framework";
 import { adaptive } from "@toss/tds-colors";
 import { Asset, FixedBottomCTA, Top } from "@toss/tds-mobile";
-import { useEffect } from "react";
 import { useMultiStep } from "../../contexts/MultiStepContext";
 import { useSurvey } from "../../contexts/SurveyContext";
+import { useBackEventListener } from "../../hooks/useBackEventListener";
 import { createScreenings } from "../../service/form";
 
 export const ScreeningOption = () => {
@@ -27,18 +26,7 @@ export const ScreeningOption = () => {
 		}
 	};
 
-	useEffect(() => {
-		const unsubscription = graniteEvent.addEventListener("backEvent", {
-			onEvent: () => {
-				goPrevScreening();
-			},
-			onError: (error) => {
-				alert(`에러가 발생했어요: ${error}`);
-			},
-		});
-
-		return unsubscription;
-	}, [goPrevScreening]);
+	useBackEventListener(goPrevScreening);
 
 	return (
 		<>
