+++ src/pages/screening/ScreeningSuccess.tsx
@@ -1,9 +1,8 @@
-import { graniteEvent } from "@apps-in-toss/web-framework";
 import { adaptive } from "@toss/tds-colors";
 import { Asset, FixedBottomCTA, Text, Top } from "@toss/tds-mobile";
-import { useEffect } from "react";
 import { useMultiStep } from "../../contexts/MultiStepContext";
 import { useSurvey } from "../../contexts/SurveyContext";
+import { useBackEventListener } from "../../hooks/useBackEventListener";
 
 export const ScreeningSuccess = () => {
 	const { handleStepChange, goPrevScreening } = useMultiStep();
@@ -11,18 +10,7 @@ export const ScreeningSuccess = () => {
 	const question = state.screening.question;
 	const selected = state.screening.answerType;
 
-	useEffect(() => {
-		const unsubscription = graniteEvent.addEventListener("backEvent", {
-			onEvent: () => {
-				goPrevScreening();
-			},
-			onError: (error) => {
-				alert(`에러가 발생했어요: ${error}`);
-			},
-		});
-
-		return unsubscription;
-	}, [goPrevScreening]);
+	useBackEventListener(goPrevScreening);
 
 	return (
 		<>
